* 第二版跟新 * 增加数字类型小数位位数限制

+++ src/main/java/org/warai/paramcheck/annotation/MinValue.java
@@ -12,7 +12,7 @@ import java.lang.annotation.*;
 @Target(ElementType.FIELD)
 public @interface MinValue {
 
-    int value();
+    double value();
 
     String msg() default ErrorMessage.BELOW_MIN_VALUE;
